@@ -1,12 +1,12 @@
 ext {
 
     android = [ // 配置信息
-                compileSdkVersion: 28,
-                buildToolsVersion: "28.0.3",
-                minSdkVersion    : 19,
-                targetSdkVersion : 28,
-                versionCode : 2,
-                versionName : "1.0.1"
+                compileSdkVersion: 29,
+                buildToolsVersion: "29.0.0",
+                minSdkVersion    : 14,
+                targetSdkVersion : 29,
+                versionCode : 3,
+                versionName : "1.0.2"
     ]
 
     // 依赖第三方库版本
