@@ -3,7 +3,7 @@ package com.henley.permissionhelper.demo;
 import android.Manifest;
 import android.app.Fragment;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
