@@ -2,7 +2,7 @@ package com.henley.permissionhelper.demo;
 
 import android.content.Context;
 import android.content.DialogInterface;
-import android.support.v7.app.AlertDialog;
+import androidx.appcompat.app.AlertDialog;
 import android.widget.Toast;
 
 import com.henley.permissionhelper.OnPermissionsCallback;
